Adds first version of krustlet based agent implementation (#18)
This is the first code drop for the agent version that has been based on the Krustlet (https://github.com/deislabs/krustlet) code. Krustlet is a virtual kubelet written in Rust and allows the user to define a provider which can act upon pods that are assigned to it.

The main part of this commit is the creation of a Krustlet provider to run processes directly on the os.

The currently implemented functionality is:

- register node with apiserver and keep the node alive
- watch for pods
- upon pod assigned:
- download package
- install package
- backoff on missing config map
- create config from config maps
- start process
- wait for process to not die during startup
- watch process for the runtime

This also adds interim command line parsing. This is not perfect yet due to limitations of the config parsing that Krustlet does.
Krustlet does not allow building a config from scratch by just specifying a few values, so we override some values by setting environment variables from our config - BUT: it is not currently possible to specify command line parameters when starting the agent, as the Krustlet will also try to parse these and doesn't know about them, so the only possible way to pass arguments is currently via exporting a config file in the CONFIG_FILE env variable and specifying all arguments in there.