Added monitoring of systemd units to notice when they fail.
During startup, every unit will be checked once every second for 10 seconds, after this it will be considered as "running".

When a unit is considered as "running" the state will be checked once every 10 seconds.

This is a simpler version of what we want to provide in a final implementation, where we listen to changes on dbus instead of polling periodically.