Adds bootstrapping of certificates and kubeconfig to agent.
If a bootstrap config is provided via the --bootstrap-file command line option and no valid config can be found in the usual places the agent will attempt to bootstrap a Kubeconfig via the configuration in the bootstrap file.

Additionally this code will check if the provided files from --server-cert-file and --server-key-file exist. If not, it will create a key pair and signing request, upload this to the apiserver and wait for the certificate to be provided by a controller.