Change check that controls whether gh actions try to publish artifacts for a pull request (#192)
We started using the action-has-permission action[1] in our workflows a while ago in order to avoid unnecessary job failures, when dependabot created pull requests that don't have access to organization secrets, or someone creates a PR from a fork (which has the same issue).

However, this didn't seem to work for forks (as tested in #177 [2]) so I tried changing the way that we check for permissions to instead simply assign the secret to an env var and then check that in the relevant actions.

This PR is to ensure actions work as intended when created from the main repo.

I believe this behavior is triggered by the way in which the action checks permissions. If I don't misread the code[3], it checks whether the user (owner of the forked repo I assume?) has permissions on the main repo. Since my owner is actually an admin in that repo that check succeeds, but github still won't make secrets available to actions run from a fork so the actual repo push then fails.

[1] https://github.com/scherermichael-oss/action-has-permission
[2] https://github.com/stackabletech/agent/pull/177
[3] https://github.com/scherermichael-oss/action-has-permission/blob/master/lib/checkPermission.js#L26