Remove restart capability from pod state
The running state is no longer responsible for deciding if a service
should be restarted. This is done by systemd. Therefore the transition
from the running state to the starting state was removed.

The service_state function was added to the SystemdManager to check if a
service is running or terminated. This function is used for the
transition from the running to the terminated state and for patching the
container state.