Move restart policy to systemd (#263)
* Set TimeoutStopSec only once

* Map restart policy to systemd service

* Remove restart capability from pod state

The running state is no longer responsible for deciding if a service should be restarted. This is done by systemd. Therefore the transition from the running state to the starting state was removed.

The service_state function was added to the SystemdManager to check if a service is running or terminated. This function is used for the transition from the running to the terminated state and for patching the container state.

* Service state "Created" added

* Waiting after systemd service startup removed

When a systemd service is started then the agent does not wait anymore ten seconds to check if it was successfully started because systemd manages restarts now and the agent cannot detect if the service is in a restart loop.

* Set RestartSec to 2 in unit files

* Start rate limiting of systemd units disabled

* Changelog updated

* Remove the waiting phase also from the docs

* Set restart count in the container status

* Set correct systemd default target